+++ ui/src/app/utils/authentication.tsx
@@ -143,6 +143,28 @@ const getAuthInstance = () => {
   return gapi.auth2.getAuthInstance();
 };
 
+export const getAccessToken = () => {
+  const tokenOverride = window.localStorage.getItem(
+    LOCAL_STORAGE_KEY_TEST_ACCESS_TOKEN
+  );
+
+  if (tokenOverride) {
+    return tokenOverride;
+  } else if (!gapi.auth2) {
+    return null;
+  } else {
+    const authResponse = gapi.auth2
+      .getAuthInstance()
+      .currentUser.get()
+      .getAuthResponse(true);
+    if (authResponse !== null) {
+      return authResponse.access_token;
+    } else {
+      return null;
+    }
+  }
+};
+
 export const hasBillingScope = () => {
   // If uses access token, assume users always have billing scope. The token generated by GenerateImpersonatedUserTokens tool sets billing
   // scope.
